notification_service/notification_service_app/serializers.py
from rest_framework import serializers
from .models import Notification, NotificationTemplate

class NotificationTemplateSerializer(serializers.ModelSerializer):
    class Meta:
        model = NotificationTemplate
        fields = '__all__'

class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = '__all__'